pyproject.toml
[tool.poetry]
name = "rndstsnd"
version = "0.1.0"
description = ""
authors = ["istimaldar <istimaldar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.3"
dynaconf = "^3.1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
